fix all in PR comments
