+++ client/src/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import todoReducer from "./todoSlice";
+import todoReducer from "@/store/todoSlice";
 import { useDispatch } from "react-redux";
 
 export const store = configureStore({
