+++ client/src/store/types.ts
@@ -1,0 +1,11 @@
+import type { Todo } from "@/api/types";
+
+export interface TodoState {
+  items: Todo[];
+  page: number;
+  limit: number;
+  total: number;
+  totalPages: number;
+  isLoading: boolean;
+  error: string | null;
+}
