+++ client/tsconfig.json
@@ -1,7 +1,17 @@
 {
-  "files": [],
-  "references": [
-    { "path": "./tsconfig.app.json" },
-    { "path": "./tsconfig.node.json" }
-  ]
+  "compilerOptions": {
+    "target": "ESNext",
+    "module": "ESNext",
+    "jsx": "react-jsx", // Поддержка React
+    "moduleResolution": "node",
+    "baseUrl": "./",
+    "paths": {
+      // Алиасы (если нужны)
+      "@/*": ["src/*"]
+    },
+    "types": ["vite/client"], // Для импортов из Vite
+    "allowSyntheticDefaultImports": true, // Разрешает дефолтные импорты
+    "esModuleInterop": true
+  },
+  "include": ["src"]
 }
